services-api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pandas
numpy
scikit-learn
tensorflow
pydantic>=2.0.0
python-dotenv
python-chess
